chore: move Axios and JSSoup to devDependencies
It seems they are only used in src/build/icons.ts, which is never called by anything.
My understanding is that this is meant to be run on its own to help with making
products. Maybe I can split this off into something else in the future – I've wanted
to make a GUI for people to do rubrics for awhile that would make PRs easier to handle,
but we'll see.

For now, the deps have been moved.